Initial commit for Angular-WinJS wrapper
This is the initial commit for our Angular-WinJS wrapper layer. This
code was ported from spankyj's original wrapper and has been updated to
be used with WinJS 4.0. We have added new directives for WinJS 3.0 and
WinJS 4.0 controls, updated the API usage for this wrapper, made several
bug fixes, and added Karma unit tests to test basic functionality for
these directives