Automate release process
You can now use a grunt task to:
  - Publish a release to GitHub
  - Publish to npm
  - Publish to bower
  - Publish to NuGet

The command is:
  grunt publish

This will print a message warning you about what is about to happen and
showing you a checklist. Once you're sure you want to proceed, you run
this command:

  grunt publish:force

which actually publishes to all of the package managers.

Version information in bower.json and Angular.WinJS.nuspec is not used.
Deleted it from bower.json. Can't delete it from nuspec so set it at
version 1.0.0 as a hint that we don't need to keep it up to date.